+++ packages/signals_flutter/lib/signals_flutter.dart
@@ -9,6 +9,8 @@ export 'src/core/signal.dart';
 export 'src/core/computed.dart';
 export 'src/ticker.dart';
 export 'src/mixins/signals.dart';
+export 'src/mixins/value_listenable.dart';
+export 'src/mixins/value_notifier.dart';
 export 'src/inherited_widget/signal.dart';
 export 'src/inherited_widget/computed.dart';
 export 'src/inherited_widget/readonly.dart';